moved instrument handling from HAL_Transmon.py to new HAL_ShimSQ.py
